2022 day06 and 07
