--- advent_of_code/2022/src/day_06.rs
@@ -1,50 +1,0 @@
-use crate::AoCData;
-
-pub struct Data<'a>(&'a str);
-
-fn helper(s: &str, size: usize) -> usize {
-    s.as_bytes()
-        .windows(size)
-        .position(|window| {
-            window
-                .iter()
-                .enumerate()
-                .all(|(idx, c)| !window[..idx].contains(c))
-        })
-        .unwrap_or(s.len())
-        + size
-}
-
-impl<'a> AoCData<'a> for Data<'a> {
-    fn try_new(input: &'a str) -> Option<Self> {
-        Some(Self(input))
-    }
-
-    fn part_1(&self) -> String {
-        helper(self.0, 4).to_string()
-    }
-
-    fn part_2(&self) -> String {
-        helper(self.0, 14).to_string()
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::utils;
-
-    #[test]
-    fn part_1() {
-        let input = utils::get_sample_input(6);
-        let data = Data::try_new(&input).unwrap();
-        assert_eq!(data.part_1(), "11");
-    }
-
-    #[test]
-    fn part_2() {
-        let input = utils::get_sample_input(6);
-        let data = Data::try_new(&input).unwrap();
-        assert_eq!(data.part_2(), "26");
-    }
-}
