--- advent_of_code/aoc/2022/inputs/day06_sample.txt
@@ -1,1 +1,0 @@
-zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw
